// src/glitchtip/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        CommitSchema: {
            /** Id */
            id: string;
            /**
             * Message
             * @default
             */
            message: string | null;
            /** Datecreated */
            dateCreated?: string | null;
            /** Authorname */
            authorName?: string | null;
            /** Authoremail */
            authorEmail?: string | null;
        };
// ...
        NameSlugProjectSchema: {
            /** Name */
            name: string;
            /** Slug */
            slug?: string | null;
        };
        /** ReleaseRepositorySchema */
        ReleaseRepositorySchema: {
            /** Id */
            id: string;
            /** Name */
            name: string;
        };
        /** ReleaseSchema */
        ReleaseSchema: {
            /** Ref */
            ref?: string | null;
            /** Datereleased */
            dateReleased: string | null;
            /** Version */
            version: string;
            /**
             * Datecreated
             * Format: date-time
             */
            dateCreated: string;
            /** Shortversion */
            shortVersion: string;
            /** Projects */
            projects: components["schemas"]["NameSlugProjectSchema"][];
            repository?: components["schemas"]["ReleaseRepositorySchema"] | null;
            /** Url */
            url?: string | null;
            /** Data */
            data?: Record<string, never>;
            /**
             * Commit Count
             * @default 0
             */
            commitCount: number;
            /**
             * Deploy Count
             * @default 0
             */
            deployCount: number;
        };
// ...
        DebugSymbolBundleSchema: {
            /** Id */
            id: string;
            /**
             * Datecreated
             * Format: date-time
             */
            dateCreated: string;
            /** Sha1 */
            sha1?: string | null;
            /** Headers */
            headers?: {
                [key: string]: string;
            } | null;
            /** Name */
            name: string;
            /**
             * Size
             * @default 0
             */
            size: number;
        };
        /** DeploySchema */
        DeploySchema: {
            /**
             * Datecreated
             * Format: date-time
             */
            dateCreated: string;
            /** Datestarted */
            dateStarted: string | null;
            /** Datefinished */
            dateFinished: string | null;
            /** ID */
            id?: number | null;
            /** Environment */
            environment: string;
            /**
             * Url
             * @default
             */
            url: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}